make a start on getting code compiling under 64bit Fix some pointer size and windows api issues
git-svn-id: https://svn.code.sf.net/p/deinterlace/code/trunk/DScaler@4600 6e12ab15-cf2f-4ffe-bf36-45892a6a5b89

+++ DScaler/OSD.cpp
@@ -214,7 +214,7 @@ void OSD_ShowText(const tstring& Text, double dSize, BOOL bPersistent, BOOL bOve
     pOSDCommand->uFlags = 0;
     pOSDCommand->dSize = dSize;
 
-    ULONG nDataLength = Text.length() + 1;
+    size_t nDataLength = Text.length() + 1;
 
     LPTSTR pszText = (LPTSTR)malloc(nDataLength);
     memcpy(pszText, Text.c_str(), nDataLength);
